Refactor Grid component and update icons; add new images for academic calendar, course browser, dining, and SIS
