+++ app/components/svgs/licenses.txt
@@ -26,3 +26,5 @@ https://www.svgrepo.com/svg/522567/info
 
 BACK CHEVRON:
 https://www.svgrepo.com/svg/522037/chevron-left
+
+https://www.svgrepo.com/svg/506814/computer
